requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
twilio==9.0.0
groq==0.11.0
google-generativeai==0.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
httpx==0.27.0
